Don't sleep before checking queues
This adds unnecessary latency to the communication, to the point where a round
trip between VXS and PX4 can be up to 40ms delayed (10ms each on incoming and
outgoing queues, on way out and way back).

This is the reason why PX4 TimeSync fails, since it has a latency requirement
of 10ms.

Also the sleeps add unnecessary jitter into the communication.